refactor(types): 카테고리 관련 타입 분리
